[spirv-writer] Start global variable output
This CL starts adding global variable output to the SPIR-V writer. This
does not handle constants or initializers at this point.

Bug: tint:5
Reviewed-on: https://dawn-review.googlesource.com/c/tint/+/17922

+++ src/ast/binding_decoration.cc
@@ -17,7 +17,7 @@
 namespace tint {
 namespace ast {
 
-BindingDecoration::BindingDecoration(size_t val) : value_(val) {}
+BindingDecoration::BindingDecoration(uint32_t val) : value_(val) {}
 
 BindingDecoration::~BindingDecoration() = default;
