+++ src/ast/binding_decoration.h
@@ -27,21 +27,21 @@ class BindingDecoration : public VariableDecoration {
  public:
   /// constructor
   /// @param value the binding value
-  explicit BindingDecoration(size_t value);
+  explicit BindingDecoration(uint32_t value);
   ~BindingDecoration() override;
 
   /// @returns true if this is a binding decoration
   bool IsBinding() const override { return true; }
 
   /// @returns the binding value
-  size_t value() const { return value_; }
+  uint32_t value() const { return value_; }
 
   /// Outputs the decoration to the given stream
   /// @param out the stream to output too
   void to_str(std::ostream& out) const override;
 
  private:
-  size_t value_;
+  uint32_t value_;
 };
 
 }  // namespace ast
