+++ src/ast/set_decoration.h
@@ -27,21 +27,21 @@ class SetDecoration : public VariableDecoration {
  public:
   /// constructor
   /// @param value the set value
-  explicit SetDecoration(size_t value);
+  explicit SetDecoration(uint32_t value);
   ~SetDecoration() override;
 
   /// @returns true if this is a set decoration
   bool IsSet() const override { return true; }
 
   /// @returns the set value
-  size_t value() const { return value_; }
+  uint32_t value() const { return value_; }
 
   /// Outputs the decoration to the given stream
   /// @param out the stream to output too
   void to_str(std::ostream& out) const override;
 
  private:
-  size_t value_;
+  uint32_t value_;
 };
 
 }  // namespace ast
